fix: do not need to enable telemetry in kubewarden-defaults.
Updates the telemetry tests removing the need to enable telemetry in the
kubewarden-defaults charts. In the latest version of the Kubewarden
controller the telemetry will be enable in the policy servers if the
telemetry is enabled in the controller. Therefore, the field to enable
this features in the kubewarden-defaults charts has been removed.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>