fix: reuse values when upgrading controller in OPTEL tests.
Reuse helm values in OpenTelemetry tests to avoid issues in the other
tests due a recent change in the controller where the finalizer used has
changed. Currently, the test downgrade the controller version causing an
issue in tests ran after the OPTEL one. Using the `--reuse-values` fix
the issue because the controller container image will not be changed.

The test to disable telemetry has been updated to use the correct values
fields to disable tracing and metrics.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>