fix: remove Otel resources before running other Otel test.
Depending the order of execution of the OpenTelemetry tests they can
fail because one test is trying to install/upgrade a resource which are
in terminating state. It's possible to see error messages like:

"create not allowed while custom resource definition is terminating"

To workaround this problem, this commit updates the test that verify the
communication with remote Otel collector to removes the resources
created during the test as well.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>